dom/tests/mochitest/notification/desktop-notification/test_notification_tag.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=782211
-->
<head>
  <title>Bug 782211</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=782211">Bug 782211</a>
<p id="display"></p>
<iframe name="sameDomain"></iframe>
<iframe name="anotherSameDomain"></iframe>
<iframe name="crossDomain"></iframe>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
<script type="text/javascript">
  const MOCK_CID = SpecialPowers.wrap(SpecialPowers.Components).ID("{dbe37e64-d9a3-402c-8d8a-0826c619f7ad}");
  const ALERTS_SERVICE_CONTRACT_ID = "@mozilla.org/alerts-service;1";

  var mockAlertsService = {
    showAlert: function(alert, alertListener) {
      notificationsCreated.push(alert.name);
      if (notificationsCreated.length == 3) {
        checkNotifications();
      }
    },

    showAlertNotification: function(imageUrl, title, text, textClickable,
                                    cookie, alertListener, name, dir,
                                    lang, data) {
      this.showAlert({ name: name });
    },

    QueryInterface: function(aIID) {
      if (SpecialPowers.wrap(aIID).equals(SpecialPowers.Ci.nsISupports) ||
          SpecialPowers.wrap(aIID).equals(SpecialPowers.Ci.nsIAlertsService)) {
        return this;
      }
      throw SpecialPowers.Components.results.NS_ERROR_NO_INTERFACE;
    },

    createInstance: function(aOuter, aIID) {
      if (aOuter != null) {
        throw SpecialPowers.Components.results.NS_ERROR_NO_AGGREGATION;
      }
      return this.QueryInterface(aIID);
    }
  };
  mockAlertsService = SpecialPowers.wrapCallbackObject(mockAlertsService);

  var notificationsCreated = [];
  function checkNotifications() {
    // notifications created by the test1 origin
    var test1notifications = [];
    // notifications created by the test2 origin
    var test2notifications = [];
    for (var i = 0; i < notificationsCreated.length; i++) {
      var notificationName = notificationsCreated[i];
      if (notificationName.indexOf("test1") !== -1) {
        test1notifications.push(notificationsCreated[i]);
      } else if (notificationName.indexOf("test2") !== -1) {
        test2notifications.push(notificationsCreated[i]);
      }
    }

    is(test1notifications.length, 2, "2 notifications should be created by test1.example.org:80 origin.");
    is(test1notifications[0], test1notifications[1], "notification names should be identical.");
    is(test2notifications.length, 1, "1 notification should be created by test2.example.org:80 origin.");

    // Register original alerts service.
    SpecialPowers.wrap(SpecialPowers.Components).
      manager.QueryInterface(SpecialPowers.Ci.nsIComponentRegistrar).
      unregisterFactory(MOCK_CID, mockAlertsService);

    SimpleTest.finish();
  }

  if (window.Notification) {
    SimpleTest.waitForExplicitFinish();

    function showNotifications() {
      SpecialPowers.wrap(SpecialPowers.Components).
        manager.QueryInterface(SpecialPowers.Ci.nsIComponentRegistrar).
        registerFactory(MOCK_CID, "alerts service", ALERTS_SERVICE_CONTRACT_ID, mockAlertsService);

      // Load two frames with the same origin that create notification with the same tag.
      // Both pages should generate notifications with the same name, and thus the second
      // notification should replace the first.
      frames["sameDomain"].location.href = "http://test1.example.org:80/tests/dom/tests/mochitest/notification/create_notification.html";
      frames["anotherSameDomain"].location.href = "http://test1.example.org:80/tests/dom/tests/mochitest/notification/create_notification.html";

      // Load a frame with a different origin that creates a notification with the same tag.
      // The notification name should be different and thus no notifications should be replaced.
      frames["crossDomain"].location.href = "http://test2.example.org:80/tests/dom/tests/mochitest/notification/create_notification.html";
    }

    SpecialPowers.pushPrefEnv({'set': [["notification.prompt.testing", true],
                                       ["notification.prompt.testing.allow", true]]},
                              showNotifications);
  } else {
    ok(true, "Notifications are not enabled on the platform.");
  }
</script>
</body>
</html>
